Refactor parser classes
