+++ src/main/java/seedu/address/logic/parser/modulelesson/EditModuleLessonCommandParser.java
@@ -1,0 +1,18 @@
+package seedu.address.logic.parser.modulelesson;
+
+import seedu.address.logic.commands.modulelesson.EditModuleLessonCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+public class EditModuleLessonCommandParser implements Parser<EditModuleLessonCommand> {
+    /**
+     * Parses {@code userInput} into a command and returns it.
+     *
+     * @param userInput
+     * @throws ParseException if {@code userInput} does not conform the expected format
+     */
+    @Override
+    public EditModuleLessonCommand parse(String userInput) throws ParseException {
+        return null;
+    }
+}
