+++ src/main/java/seedu/address/logic/parser/modulelesson/FindModuleLessonCommandParser.java
@@ -1,0 +1,18 @@
+package seedu.address.logic.parser.modulelesson;
+
+import seedu.address.logic.commands.modulelesson.FindModuleLessonCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+public class FindModuleLessonCommandParser implements Parser<FindModuleLessonCommand> {
+    /**
+     * Parses {@code userInput} into a command and returns it.
+     *
+     * @param userInput
+     * @throws ParseException if {@code userInput} does not conform the expected format
+     */
+    @Override
+    public FindModuleLessonCommand parse(String userInput) throws ParseException {
+        return null;
+    }
+}
