+++ src/main/java/seedu/address/logic/parser/modulelesson/RemarkModuleLessonCommandParser.java
@@ -1,0 +1,18 @@
+package seedu.address.logic.parser.modulelesson;
+
+import seedu.address.logic.commands.modulelesson.RemarkModuleLessonCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+public class RemarkModuleLessonCommandParser implements Parser<RemarkModuleLessonCommand> {
+    /**
+     * Parses {@code userInput} into a command and returns it.
+     *
+     * @param userInput
+     * @throws ParseException if {@code userInput} does not conform the expected format
+     */
+    @Override
+    public RemarkModuleLessonCommand parse(String userInput) throws ParseException {
+        return null;
+    }
+}
