+++ src/test/java/seedu/address/logic/parser/AddPersonCommandParserTest.java
@@ -36,6 +36,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.person.AddPersonCommand;
+import seedu.address.logic.parser.person.AddCommandParser;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.ModuleCode;
 import seedu.address.model.person.Name;
