+++ src/test/java/seedu/address/logic/parser/FindPersonCommandParserTest.java
@@ -10,6 +10,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.person.FindPersonCommand;
+import seedu.address.logic.parser.person.FindCommandParser;
 import seedu.address.model.person.ModuleCodesContainsKeywordsPredicate;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
